양도가격 최대 999,000워으로 설정 구현

+++ src/main/java/com/springbootfinal/app/domain/transfer/TransferDto.java
@@ -20,7 +20,9 @@ public class TransferDto {
     private Long sellerUserNo;
     private Long buyerUserNo;
     private Long reservationNo;
-    private BigDecimal transferPrice;
+
+    private Long transferPrice;
+
     private String status;
     private LocalDateTime createdAt;
     private String transferContent;
